app/requirements.txt
alembic==1.9.1
anyio==3.6.2
click==8.1.3
fastapi==0.89.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
pydantic==1.10.4
python-dotenv==0.21.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
typing==3.7.4.3
typing_extensions==4.4.0
uvicorn==0.20.0
